kubeflow_women_clothing_ecom_reviews/pipeline/components/data_preparation/requirements.txt
pyyaml==5.4.1
pandas==1.1.5
nltk==3.5
# for local testing
# kfp==1.4.0